pyproject.toml
[project]
name = "ai-ice-breaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "gradio>=6.8.0",
    "groq>=0.37.1",
    "huggingface-hub>=0.36.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "llama-hub>=0.0.79.post1",
    "llama-index>=0.14.15",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-groq>=0.4.1",
    "llama-index-readers-web>=0.5.6",
    "llama-index-vector-stores-chroma>=0.5.5",
    "notebook>=7.5.4",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "wget>=3.2",
]
